+++ cw_config_ecs/main.tf
@@ -1,5 +1,5 @@
 module "ssm_param_cw_config" {
-  source = "../ssm_parameter_plain"
+  source = "../ssm/parameter_plain"
   param_map = {
     cloudwatch_agent_config_ecs_cpu = {
       insecure_value = jsonencode(local.cw_config_cpu)