+++ sqs_dead_letter_queue/main.tf
@@ -1,5 +1,0 @@
-module "this_queue" {
-  source    = "../sqs_queue"
-  queue_map = local.queue_map
-  std_map   = var.std_map
-}